Refactor Z3 timeout (#2148)
* Check Z3 timeout for allowed values

* Safely convert timeout to int

Z3 takes the timeout in milliseconds as unsigned int. So we can savely
store timeout in seconds as (signed) int.

+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/smt/PreproSolverContext.scala
@@ -210,7 +210,7 @@ class PreproSolverContext(context: SolverContext) extends SolverContext {
    * @return
    *   Some(result), if no timeout happened; otherwise, None
    */
-  override def satOrTimeout(timeoutSec: Long): Option[Boolean] = context.satOrTimeout(timeoutSec)
+  override def satOrTimeout(timeoutSec: Int): Option[Boolean] = context.satOrTimeout(timeoutSec)
 
   /**
    * Register an SMT listener
